@@ -1,17 +1,6 @@
 import numba
 import numpy as np
 
-__all__ = [
-    "potbesldho",
-    "potbeslsho",
-    "potbesldv",
-    "potbeslsv",
-    "disbesldho",
-    "disbeslsho",
-    "disbesldv",
-    "disbeslsv",
-]
-
 
 def initialize():
     pass
@@ -443,7 +432,7 @@ def IntegralF(zin, z1in, z2in, Lin, labda, order, Rconv, lstype):
     zbar = np.conj(z)
 
     # Coefficients gamma(n,m), Eq. 21
-    # Store coefficents in matrix.
+    # Store coefficients in matrix.
     for n in range(0, NTERMS + 1):
         czmzbarp[n] = (z - zbar) ** n
 
@@ -526,7 +515,7 @@ def IntegralG(zin, z1in, z2in, Lin, labda, order, Rconv, lstype):
     zbar = np.conj(z)
 
     # Coefficients gamma(n,m), Eq. 21
-    # Store coefficents in matrix.
+    # Store coefficients in matrix.
     for n in range(0, NTERMS + 1):
         czmzbarp[n] = (z - zbar) ** n
 
