@@ -1,3 +1,15 @@
+"""1D inhomogeneities.
+
+Cross-section inhomogeneity elements. Used to build cross-section models with
+varying aquifer properties or top boundary conditions along a cross-section.
+
+Example::
+
+    XsectionMaq(ml, x1=-np.inf, x2=0, kaq=[10, 1], z=[10, 5, 0], c=[100])
+    XsectionMaq(ml, x1=0, x2=np.inf, kaq=[1, 10], z=[10, 5, 0], c=[50])
+
+"""
+
 import inspect  # user for storing the input
 import warnings
 
@@ -10,7 +22,7 @@ from timml.constant import ConstantStar
 from timml.linesink1d import FluxDiffLineSink1D, HeadDiffLineSink1D
 from timml.stripareasink import XsectionAreaSinkInhom
 
-__all__ = ["XsectionMaq", "Xsection3D"]
+__all__ = ["XsectionMaq", "Xsection3D", "Xsection"]
 
 
 class Xsection(AquiferData):
@@ -41,6 +53,7 @@ class Xsection(AquiferData):
         infiltration rate, only read if topboundary='conf'
     name : string, optional
         name of the inhomogeneity
+
     """
 
     tiny = 1e-12
@@ -127,7 +140,9 @@ class Xsection(AquiferData):
             c = ConstantStar(self.model, self.hstar, aq=aqin)
             c.inhomelement = True
 
-    def plot(self, ax=None, labels=False, params=False, names=False, **kwargs):
+    def plot(
+        self, ax=None, labels=False, params=False, names=False, fmt=None, **kwargs
+    ):
         """Plot the cross-section.
 
         Parameters
@@ -140,6 +155,8 @@ class Xsection(AquiferData):
             If True, add parameter labels.
         names : bool, optional
             If True, add inhomogeneity names.
+        fmt : str
+            format string for parameters
         """
         if ax is None:
             _, ax = plt.subplots(1, 1, figsize=(8, 4))
@@ -167,6 +184,9 @@ class Xsection(AquiferData):
 
         r = x2 - x1
         r0 = x1
+
+        if fmt is None:
+            fmt = ""
 
         if labels or params:
             lli = 1 if self.ltype[0] == "a" else 0
@@ -203,7 +223,7 @@ class Xsection(AquiferData):
                         va="center",
                     )
                 if params:
-                    paramtxt = f"$c$ = {self.c[lli]:.1f}"
+                    paramtxt = f"$c$ = {self.c[lli]:{fmt}}"
                     ax.text(
                         r0 + 0.75 * r if labels else r0 + 0.5 * r,
                         np.mean(self.z[i : i + 2]),
@@ -223,7 +243,7 @@ class Xsection(AquiferData):
                     va="center",
                 )
             if params and self.ltype[i] == "a":
-                paramtxt = f"$k_h$ = {self.kaq[aqi]:.1f}"
+                paramtxt = f"$k_h$ = {self.kaq[aqi]:{fmt}}"
                 ax.text(
                     r0 + 0.75 * r if labels else r0 + 0.5 * r,
                     np.mean(self.z[i : i + 2]),
