@@ -1,3 +1,12 @@
+"""Line-doublet elements.
+
+Implements line doublets used to model barriers.
+
+Example::
+
+    ImpLineDoublet(ml, x1=-10, y1=0, x2=10, y2=0, layers=0)
+"""
+
 import inspect  # Used for storing the input
 
 import matplotlib.pyplot as plt
@@ -415,7 +424,7 @@ class LineDoubletStringBase(Element):
 
 
 class ImpLineDoubletString(LineDoubletStringBase, DisvecEquation):
-    """Create a string of impermeable wall segements consisting of line-doublets.
+    """Create a string of impermeable wall segments consisting of line-doublets.
 
     Parameters
     ----------
@@ -423,7 +432,7 @@ class ImpLineDoubletString(LineDoubletStringBase, DisvecEquation):
         Model to which the element is added
     xy : array or list
         list or array of (x,y) pairs of coordinates of end-points of
-        the segements in the string
+        the segments in the string
     layers : scalar, list or array
         layer(s) in which element is placed
         if scalar: element is placed in this layer
@@ -466,7 +475,7 @@ class ImpLineDoubletString(LineDoubletStringBase, DisvecEquation):
 
 
 class LeakyLineDoubletString(LineDoubletStringBase, LeakyWallEquation):
-    """Create a string of leaky wall segements consisting of line-doublets.
+    """Create a string of leaky wall segments consisting of line-doublets.
 
     Parameters
     ----------
@@ -474,7 +483,7 @@ class LeakyLineDoubletString(LineDoubletStringBase, LeakyWallEquation):
         Model to which the element is added
     xy : array or list
         list or array of (x,y) pairs of coordinates of end-points of
-        the segements in the string
+        the segments in the string
     res : scalar
         resistance of leaky wall
     layers : scalar, list or array
