@@ -1,4 +1,13 @@
-from typing import Optional
+"""Plot helpers for TimML.
+
+Provides top-view, contours, and tracing visualization functions.
+
+Example::
+
+    ml.plots.topview()
+"""
+
+from typing import Literal, Optional
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -9,8 +18,16 @@ from timml.trace import timtraceline
 
 plt.rcParams["contour.negative_linestyle"] = "solid"
 
+__all__ = ["PlotTim"]
+
 
 class PlotTim:
+    """Plotting functionality for TimML models.
+
+    Provides methods for visualizing model layouts, contours, pathlines,
+    and other model results.
+    """
+
     def __init__(self, ml):
         self._ml = ml
 
@@ -81,13 +98,13 @@ class PlotTim:
                 ax1 = fig.axes[0]
                 ax2 = fig.axes[1]
             elif orientation[:3] == "hor":
-                fig = plt.gcf()
-                ax1 = fig.axes[0]
+                ax1 = plt.gca()
                 ax2 = None
+                fig = ax1.figure
             elif orientation[:3] == "ver":
-                fig = plt.gcf()
                 ax1 = None
-                ax2 = fig.axes[0]
+                ax2 = plt.gca()
+                fig = ax2.figure
         if ax1 is not None:
             plt.sca(ax1)
             for e in self._ml.elementlist:
@@ -121,6 +138,9 @@ class PlotTim:
         labels=True,
         params=False,
         ax=None,
+        fmt=None,
+        units=None,
+        horizontal_axis: Literal["x", "y", "s"] = "s",
     ):
         """Plot cross-section of model.
 
@@ -138,6 +158,14 @@ class PlotTim:
             add parameter values to plot
         ax : matplotlib.Axes, optional
             axes to plot on, default is None which creates a new figure
+        fmt : str, optional
+            format string for parameter values, e.g. '.2f' for 2 decimals
+        units : dict, optional
+            dictionary with units for parameters, e.g. {'k': 'm/d', 'c': 'd'}
+        horizontal_axis : str
+            's' for distance along cross-section on x-axis (default)
+            'x' for using x-coordinates on x-axis
+            'y' for using y-coordinates on x-axis
 
         Returns
         -------
@@ -156,32 +184,41 @@ class PlotTim:
             else:
                 x1, x2 = ax.get_xlim()
             for inhom in self._ml.aq.inhomlist:
-                inhom.plot(ax=ax, labels=labels, params=params, x1=x1, x2=x2)
+                inhom.plot(ax=ax, labels=labels, params=params, x1=x1, x2=x2, fmt=fmt)
             ax.set_xlim(x1, x2)
             ax.set_ylabel("elevation")
             ax.set_xlabel("x")
             return ax
 
+        if fmt is None:
+            fmt = ""
+
         # else get cross-section line
         if xy is not None:
             (x0, y0), (x1, y1) = xy
             r = np.sqrt((x1 - x0) ** 2 + (y1 - y0) ** 2)
-            if y0 == 0 and y1 == 0:
-                r0 = x0
-                ax.set_xlim(x0, x1)
-            elif x0 == 0 and x1 == 0:
-                ax.set_ylim(y0, y1)
-                r0 = y0
-            else:
+            if horizontal_axis == "s":
                 ax.set_xlim(0, r)
                 r0 = 0.0
+            elif horizontal_axis == "x":
+                ax.set_xlim(np.min([x0, x1]), np.max([x0, x1]))
+                r0 = np.min([x0, x1])
+            elif horizontal_axis == "y":
+                ax.set_xlim(np.min([y0, y1]), np.max([y0, y1]))
+                r0 = np.min([y0, y1])
+            else:
+                raise ValueError("horizontal_axis must be 'x', 'y', or 's'")
         else:
             r0 = 0.0
             r = 1.0
+            ax.set_xticks([])
 
         # get values for layer and aquifer numbering
         if labels:
-            lli = 1 if self._ml.aq.ltype[0] == "a" else 0
+            if self._ml.name == "Model":
+                lli = 0
+            else:
+                lli = 1 if self._ml.aq.ltype[0] == "a" else 0
             aqi = 0
         else:
             lli = None
@@ -205,15 +242,24 @@ class PlotTim:
                         va="center",
                     )
                 if params:
+                    if units is not None:
+                        unitstr = f" {units['c']}" if "c" in units else ""
+                    else:
+                        unitstr = ""
                     ax.text(
                         r0 + 0.75 * r if labels else r0 + 0.5 * r,
                         np.mean(self._ml.aq.z[i : i + 2]),
-                        (f"$c$ = {self._ml.aq.c[lli]}"),
+                        (f"$c$ = {self._ml.aq.c[lli]:{fmt}}" + unitstr),
                         ha="center",
                         va="center",
                     )
                 if labels or params:
                     lli += 1
+
+            # for Model class, k_h and c have to be supplied always,
+            # even for zero-thickness aquitards.
+            if self._ml.name == "Model" and aqi % 2 == 0:
+                lli += 1
 
             # aquifers
             if labels and self._ml.aq.ltype[i] == "a":
@@ -225,10 +271,11 @@ class PlotTim:
                     va="center",
                 )
             if params and self._ml.aq.ltype[i] == "a":
-                if aqi == 0:
-                    paramtxt = f"$k_h$ = {self._ml.aq.kaq[aqi]}"
-                if self._ml.name == "Model3D":
-                    paramtxt += f", $k_z/k_h$ = {self._ml.aq.kzoverkh[aqi]:.2f}"
+                if units is not None:
+                    unitstr = f" {units['k']}" if "k" in units else ""
+                else:
+                    unitstr = ""
+                paramtxt = f"$k_h$ = {self._ml.aq.kaq[aqi]:{fmt}}" + unitstr
                 ax.text(
                     r0 + 0.75 * r if labels else r0 + 0.5 * r,
                     np.mean(self._ml.aq.z[i : i + 2]),
@@ -356,6 +403,7 @@ class PlotTim:
         newfig=True,
         figsize=None,
         layout=True,
+        horizontal_axis: Literal["x", "y", "s"] = "s",
     ):
         """Head contour plot in vertical cross-section.
 
@@ -384,18 +432,28 @@ class PlotTim:
             size of figure
         layout : boolean
             plot layout if True
+        horizontal_axis : str, optional
+            's' for distance along cross-section on x-axis (default)
+            'x' for using x-coordinates on x-axis
+            'y' for using y-coordinates on x-axis
 
         Returns
         -------
         cs : contour set
         """
         x1, x2, y1, y2 = win
-        h = self._ml.headalongline(
-            np.linspace(x1 + nudge, x2 - nudge, n),
-            np.linspace(y1 + nudge, y2 - nudge, n),
-        )
-        L = np.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
-        xg = np.linspace(0, L, n)
+        xg = np.linspace(x1 + nudge, x2 - nudge, n)
+        yg = np.linspace(y1 + nudge, y2 - nudge, n)
+        h = self._ml.headalongline(xg, yg)
+        if horizontal_axis == "x":
+            xg = xg
+        elif horizontal_axis == "y":
+            xg = yg
+        elif horizontal_axis == "s":
+            L = np.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
+            xg = np.linspace(0, L, n)
+        else:
+            raise ValueError("horizontal_axis must be 'x', 'y', or 's'")
         if vinterp:
             zg = 0.5 * (self._ml.aq.zaqbot + self._ml.aq.zaqtop)
             zg = np.hstack((self._ml.aq.zaqtop[0], zg, self._ml.aq.zaqbot[-1]))
@@ -409,7 +467,12 @@ class PlotTim:
         if newfig:
             _, ax = plt.subplots(figsize=figsize)
         if layout:
-            self.xsection(xy=[(x1, y1), (x2, y2)], labels=False, ax=ax)
+            self.xsection(
+                xy=[(x1, y1), (x2, y2)],
+                labels=False,
+                ax=ax,
+                horizontal_axis=horizontal_axis,
+            )
         cs = ax.contour(xg, zg, h, levels, colors=color)
         if labels:
             fmt = "%1." + str(decimals) + "f"
@@ -471,7 +534,7 @@ class PlotTim:
             return traces if True
         metadata: boolean
             if False, return list of xyzt arrays
-            if True, return list of result dicionaries
+            if True, return list of result dictionaries
 
         Returns
         -------
