@@ -1,9 +1,20 @@
+"""Cross-section area-sink element.
+
+Models uniform infiltration over along a strip (cross-section).
+
+Example::
+
+    XsectionAreaSink(ml, xleft=-50, xright=50, N=0.001, layer=0)
+"""
+
 import warnings
 
 import matplotlib.pyplot as plt
 import numpy as np
 
 from timml.element import Element
+
+__all__ = ["XsectionAreaSinkInhom", "XsectionAreaSink"]
 
 
 class XsectionAreaSinkInhom(Element):
@@ -22,6 +33,7 @@ class XsectionAreaSinkInhom(Element):
         left boundary of inhomogeneity (may not be -np.inf)
     xright : float
         right boundary of inhomogeneity (may not be np.inf)
+
     """
 
     def __init__(
@@ -115,6 +127,11 @@ class XsectionAreaSinkInhom(Element):
 
 
 class XsectionAreaSink(Element):
+    """Cross-section area-sink for testing purposes only.
+
+    Use XsectionMaq or Xsection3D with 'N' parameter instead for production code.
+    """
+
     def __init__(
         self,
         model,
@@ -127,7 +144,7 @@ class XsectionAreaSink(Element):
     ):
         warnings.warn(
             "XsectionAreaSink is only for testing purposes. It is recommended to add "
-            "infiltration through XsectionMaq or Xsection3D and specifing 'N'.",
+            "infiltration through XsectionMaq or Xsection3D and specifying 'N'.",
             DeprecationWarning,
             stacklevel=2,
         )
