@@ -1,7 +1,9 @@
+"""Version information for TimML."""
+
 from importlib import metadata
 from platform import python_version
 
-__version__ = "6.7.0"
+__version__ = "6.8.0"
 
 
 def show_versions() -> None:
